+++ tests/test_wafv2.py
@@ -1,8 +1,19 @@
 import unittest
 
 from troposphere.wafv2 import (
+    AndStatement,
+    ByteMatchStatement,
     CustomResponseBody,
+    FieldToMatch,
+    NotStatement,
+    Statement,
+    TextTransformation,
+    VisibilityConfig,
+    WebACLRule,
+    XssMatchStatement,
     validate_custom_response_bodies,
+    validate_statement,
+    validate_statements,
     wafv2_custom_body_response_content,
     wafv2_custom_body_response_content_type,
 )
@@ -40,6 +51,69 @@ class TestWafV2CustomBodies(unittest.TestCase):
         for s in ["", "APPLICATION", "HTML", "TEXT"]:
             with self.assertRaises(ValueError):
                 wafv2_custom_body_response_content_type(s)
+
+    def test_statement_validator(self):
+        validate_statement(Statement())
+        with self.assertRaises(TypeError):
+            validate_statement("foo")
+        with self.assertRaises(TypeError):
+            validate_statement(10)
+
+    def test_statements_validator(self):
+        validate_statements([Statement(), Statement()])
+        with self.assertRaises(TypeError):
+            validate_statements(
+                [
+                    Statement(),
+                ]
+            )
+        with self.assertRaises(TypeError):
+            validate_statements("foo")
+        with self.assertRaises(TypeError):
+            validate_statements(["foo", "bar"])
+
+    def test_wafv2_snippet(self):
+        web_acl_rule = WebACLRule(
+            "WebACLRule",
+            Name="XSSprotection",
+            Priority=0,
+            VisibilityConfig=VisibilityConfig(
+                SampledRequestsEnabled=True,
+                CloudWatchMetricsEnabled=True,
+                MetricName="XSSprotection",
+            ),
+            Statement=Statement(
+                AndStatement=AndStatement(
+                    Statements=[
+                        Statement(
+                            XssMatchStatement=XssMatchStatement(
+                                FieldToMatch=FieldToMatch(UriPath={}),
+                                TextTransformations=[
+                                    TextTransformation(Type="URL_DECODE", Priority=0),
+                                ],
+                            ),
+                        ),
+                        Statement(
+                            NotStatement=NotStatement(
+                                Statement=Statement(
+                                    ByteMatchStatement=ByteMatchStatement(
+                                        FieldToMatch=FieldToMatch(UriPath={}),
+                                        PositionalConstraint="CONTAINS",
+                                        SearchString="xxxx",
+                                        TextTransformations=[
+                                            TextTransformation(
+                                                Type="LOWERCASE", Priority=0
+                                            ),
+                                        ],
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ],
+                ),
+            ),
+        )
+        web_acl_rule.to_dict()
 
 
 if __name__ == "__main__":